init rdoc branch
