+++ spec/task_runner_spec.rb
@@ -5,6 +5,28 @@ describe WaxTasks::TaskRunner do
     WaxTasks::Test.reset
   end
 
+  let(:multi_collection_runner){
+    opts = {
+      collections: {
+        'c1' => {
+          'source' => 'valid.csv',
+          'layout' => 'default.html',
+          'lunr_index' =>{
+            'fields' => ['gambrel', 'indescribable' ,'blasphemous', 'furtive']
+          }
+        },
+        'c2' => {
+          'source' => 'valid.csv',
+          'layout' => 'default.html',
+          'lunr_index' =>{
+            'fields' => ['gambrel', 'indescribable' ,'blasphemous', 'furtive']
+          }
+        }
+      }
+    }
+    WaxTasks::TaskRunner.new.override(opts)
+  }
+
   describe '.new' do
     it 'initializes without errors' do
       expect(task_runner).to be_an_instance_of(WaxTasks::TaskRunner)
@@ -67,12 +89,23 @@ describe WaxTasks::TaskRunner do
       index = File.read(index_path).remove_yaml
       File.open(index_path, 'w') { |f| f.write(index) }
       expect { WaxTasks::Utils.validate_json(index_path) }.to_not raise_error
+      expect(WaxTasks::Utils.validate_json(index_path)).not_to be_empty
     end
 
     context 'when generate_ui=true' do
       it 'generates a default ui' do
         task_runner.lunr(generate_ui=true)
         expect(File).to exist(ui_path)
+      end
+    end
+
+    context 'when given multiple collections' do
+      it 'generates an index including each collection' do
+        quiet_stdout { multi_collection_runner.pagemaster(['c1', 'c2']) }
+        lunr_collections = WaxTasks::Utils.get_lunr_collections(multi_collection_runner.site)
+        lunr_collections.map! { |name| WaxTasks::LunrCollection.new(name, multi_collection_runner.site) }
+        index = WaxTasks::LunrIndex.new(lunr_collections).to_s.remove_yaml
+        expect(JSON.load(index).length).to eq(6)
       end
     end
   end
@@ -92,7 +125,6 @@ describe WaxTasks::TaskRunner do
   describe '.js_package' do
     it 'creates a package.json file' do
       package = task_runner.js_package
-      puts package
     end
   end
 end
